chore(login flow): Refactor ButtonIceLoadingIndicator to IceLoadingIndicator
