+++ lib/app/features/auth/views/pages/sign_up_passkey/sign_up_passkey_form.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:ice/app/components/button/button.dart';
+import 'package:ice/app/components/progress_bar/ice_loading_indicator.dart';
 import 'package:ice/app/extensions/extensions.dart';
 import 'package:ice/app/features/auth/data/models/auth_state.dart';
 import 'package:ice/app/features/auth/providers/auth_provider.dart';
@@ -27,7 +28,7 @@ class SignUpPasskeyForm extends HookConsumerWidget {
           Button(
             disabled: authState is AuthenticationLoading,
             trailingIcon: authState is AuthenticationLoading
-                ? const ButtonIceLoadingIndicator()
+                ? const IceLoadingIndicator()
                 : const SizedBox.shrink(),
             onPressed: () {
               if (formKey.value.currentState!.validate()) {
